titanic/requirements.txt
kaggle==1.6.8
python-dotenv==1.0.1